Backend-python/config/db.py
import os
from motor.motor_asyncio import AsyncIOMotorClient
from dotenv import load_dotenv

load_dotenv()  # Load environment variables from .env file

MONGO_URI = os.getenv("MONGO_URI", "mongodb://localhost:27017/REN")
client = AsyncIOMotorClient(MONGO_URI)
db = client.get_default_database()
