Backend-python/requirements.txt
fastapi
uvicorn
motor
python-dotenv
google-genai
pydantic
transformers
torch
pymongo